refactor(deployment): move live lease helpers out of reclamationUtils
isLeaseLive and LIVE_LEASE_STATES are lease state, not reclamation.
Billing and the other live-lease callers now import them from leaseUtils.
