+++ apps/deploy-web/src/components/deployments/DeploymentListRow.tsx
@@ -34,9 +34,10 @@ import { useRedeploy } from "@src/hooks/useRedeploy/useRedeploy";
 import { useAllLeases, useLeaseStatus } from "@src/queries/useLeaseQuery";
 import type { LeaseDto, NamedDeploymentDto } from "@src/types/deployment";
 import type { ApiProviderList } from "@src/types/provider";
+import { isLeaseLive } from "@src/utils/leaseUtils";
 import { udenomToDenom } from "@src/utils/mathHelpers";
 import { getTimeLeft } from "@src/utils/priceUtils";
-import { getLeaseCloseReasonLabel, getReclamationDeadline, isLeaseLive, isProviderReclaimed, isReclaiming } from "@src/utils/reclamationUtils";
+import { getLeaseCloseReasonLabel, getReclamationDeadline, isProviderReclaimed, isReclaiming } from "@src/utils/reclamationUtils";
 import { TransactionMessageData } from "@src/utils/TransactionMessageData";
 import { UrlService } from "@src/utils/urlUtils";
 import { TrialDeploymentBadge } from "../shared";
