+++ apps/deploy-web/src/components/home/YourAccount/YourAccount.tsx
@@ -9,8 +9,8 @@ import type { WalletBalance } from "@src/hooks/useWalletBalance";
 import sdlStore from "@src/store/sdlStore";
 import type { DeploymentDto, LeaseDto } from "@src/types/deployment";
 import type { ApiProviderList } from "@src/types/provider";
+import { isLeaseLive } from "@src/utils/leaseUtils";
 import { getAvgCostPerMonth, getLeasesCostPerBlockUsd } from "@src/utils/priceUtils";
-import { isLeaseLive } from "@src/utils/reclamationUtils";
 import { bytesToShrink } from "@src/utils/unitUtils";
 import { AccountHeader } from "../AccountHeader";
 import { AccountStatsCards } from "../AccountStatsCards/AccountStatsCards";
